useState - Define user entered values as a state
Created a single useState with an object for all input fields. Three different change handlers are triggered on each keystroke (onChange) which fires a state-updating function for that particular state, which fires a callback function that receives the previous value to modify the particular state that triggered the state-updating function.