Achieved 2-way binding by setting input field's value property
The input field's value property is set equal to useState's managed data pointer (which points to an object [3 pieces of data]). The form's onSubmit event triggers a submitHandler which outputs the data to console, then trigger's onState's state-updating function setUserInput while passing in a new object with the managed data set to an empty string to clear the input fields.