Created Expense Filter and stored selection as a state in Expenses.js
Created ExpensesFilter.js to filter expenses by year. This component was a child component of Expense.js so the value selected by the user was passed to the parent component via a function parameter as a prop. The value was stored as a state in Expense.js and passed back into ExpensesFilter.js to create a "controlled component".