Update expense array when new expenses is added
Used useState with an initial value of the dummy data. The addExpenseHandler which is fired via the onAddExpense prop inside of NewExpense.js, which is fired via the onSaveExpenseData prop inside of ExpenseForm.js (whenever new data is submitted) fires our state-updating function setExpenses which has a callback function that takes the prevExpenses as a parameter returns a new array with the new item added to it.