Rendered ExpenseForm.js conditionally
ExpenseForm.js is nested in NewExpense.js. I added a button next to the ExpenseForm.js component that conditionally renders based on a new useState. Created a new useState with a default value of false. The button's onClick listener triggers the useState's state-changing function to change the value to true. The nested button and the nested form are conditionally rendered via an inline condition. A cancel button was added to the ExpenseForm.js component that has an onClick listener that fires an onCancel prop that was passed to ExpenseForm.js. The onCancel prop fires a stopEditingHandler which fires the SetIsEditing state-changing function to false, which hides the form. I also added the state-changing function to our form submit (saveExpenseDataHandler) to change the state to false when the form is submitted.