Created necessary bar chart components and established props
Created Chart.js with a dynamic number of nested ChartBars derived by mapping datapoints that will be passed in via a dataPoints prop. Each ChartBar has data passed into it via props including a key for efficient rendering, a value, a maxValue relative to all rendered datapoints, and a label.