Created ExpensesChart.js which is a shell for Chart.js
On ExpensesChart.js we created a chartDataPoints array full of objects which are our data points for each month. We extracted the necessary data for this array using props, then passed the array as a prop into Chart.js.

In Chart.js we extracted an array of values from the chartDataPoints object and extracted the values as standalone arguments for a max method which gives us the max value.

In the Expenses.js file, we import ExpensesChart.js and nest it between the filter and the list. We close the loop by passing the filteredExpenses as a prop into our ExpensesChart.js where it is used to create the chartDataPoints array.